Incorporated code from "The Annotated Encoder Decoder" project. https://bastings.github.io/annotated_encoder_decoder/ implementing the Bahdenau et.al encoder, such that it can be tested in the same notebook base as the transformer model, and the two models can be compared for performance.
A lot of work went into adding example length information in the collate
function, and also adapting the source padding size of the batches produced by
the collate function to be actually determined by the shortest source side
amongst the examples, rather than a fixed size. With the fixed size, it gave
problems with mismatched tensor sizes somewhere in the model.

TODO: It remains to be seen if the current adaptations also still let the
transformer model run, or more adaptations have to be made e.g. to the collate
function, to fix that (perhaps it expects the fixed-dim batches for example).

	modified:   the_annotated_transformer_plus_other_models.py